Split meter into components
